Isolate desk connector context helpers
