+++ apps/ui/src/ui-desks/connectors/utils.ts
@@ -1,21 +1,18 @@
-import { RichTextData } from '@wordpress/rich-text';
-import { sortByIndex, useValue, type Editor, type TLShape, type TLShapeId } from 'tldraw';
+import { sortByIndex, type Editor, type TLShape, type TLShapeId } from 'tldraw';
+import {
+	getDeskWidgetConnectionLabel,
+	getDeskWidgetConnectionPillBg,
+	getDeskWidgetConnectionTitle,
+	type DeskWidgetConnectionTarget,
+} from '@/ui-desks/connectors/context';
 import {
 	canvasShapeToDeskWidget,
 	isDeskConnectorCanvasShape,
 } from '@/ui-desks/desk/tldraw-adapter';
 import { getWidgetDropHandler } from '@/ui-desks/widget-actions/drop-handlers';
-import { NOTE_WIDGET_TYPE, type NoteTone } from '@/ui-desks/widgets/note/types';
-import type { DeskConfig } from '@/ui-desks/desk/types';
 import type { DeskWidget, WidgetDropHandler } from '@/ui-desks/widgets/types';
 
-export interface DeskWidgetConnectionTarget {
-	shapeId: TLShapeId;
-	widget: DeskWidget;
-	label: string;
-	title: string;
-	pillBg?: string;
-}
+export type { DeskWidgetConnectionTarget };
 
 export interface SelectedDeskConnectorToolbarItem {
 	shapeId: TLShapeId;
@@ -51,88 +48,6 @@ export function getOutgoingWidgetConnections(
 		} );
 	}
 	return targets;
-}
-
-export function useIncomingWidgetConnections(
-	editor: Editor,
-	targetShapeId: TLShapeId | undefined
-): DeskWidgetConnectionTarget[] {
-	return useValue(
-		`incoming-widget-connections:${ targetShapeId ?? 'none' }`,
-		() => ( targetShapeId ? getIncomingWidgetConnections( editor, targetShapeId ) : [] ),
-		[ editor, targetShapeId ]
-	);
-}
-
-export function getIncomingWidgetConnections(
-	editor: Editor,
-	targetShapeId: TLShapeId
-): DeskWidgetConnectionTarget[] {
-	const sources: DeskWidgetConnectionTarget[] = [];
-	for ( const binding of editor.getBindingsToShape( targetShapeId, 'arrow' ) ) {
-		if ( getArrowBindingTerminal( binding.props ) !== 'end' ) {
-			continue;
-		}
-
-		const connectorShape = editor.getShape( binding.fromId );
-		if ( ! isDeskConnectorCanvasShape( connectorShape ) ) {
-			continue;
-		}
-
-		const startBinding = editor
-			.getBindingsFromShape( connectorShape.id, 'arrow' )
-			.find( ( candidate ) => getArrowBindingTerminal( candidate.props ) === 'start' );
-		if ( ! startBinding ) {
-			continue;
-		}
-
-		const sourceShape = editor.getShape( startBinding.toId );
-		const widget = sourceShape ? canvasShapeToDeskWidget( sourceShape ) : null;
-		if ( ! sourceShape || ! widget ) {
-			continue;
-		}
-
-		const label = getDeskWidgetConnectionLabel( widget );
-		sources.push( {
-			shapeId: sourceShape.id,
-			widget,
-			label,
-			title: getDeskWidgetConnectionTitle( widget, label ),
-			pillBg: getDeskWidgetConnectionPillBg( widget ),
-		} );
-	}
-	return sources;
-}
-
-export function appendIncomingConnectedWidgets(
-	widgets: DeskWidget[],
-	deskConfig: DeskConfig | null | undefined
-): DeskWidget[] {
-	if ( widgets.length === 0 || ! deskConfig?.connectors?.length ) {
-		return widgets;
-	}
-
-	const widgetsById = new Map( deskConfig.widgets.map( ( widget ) => [ widget.id, widget ] ) );
-	const selectedIds = new Set( widgets.map( ( widget ) => widget.id ) );
-	const output = [ ...widgets ];
-
-	for ( const widget of widgets ) {
-		for ( const connector of deskConfig.connectors ) {
-			if ( connector.to.widgetId !== widget.id || selectedIds.has( connector.from.widgetId ) ) {
-				continue;
-			}
-
-			const sourceWidget = widgetsById.get( connector.from.widgetId );
-			if ( ! sourceWidget ) {
-				continue;
-			}
-
-			selectedIds.add( sourceWidget.id );
-			output.push( sourceWidget );
-		}
-	}
-
-	return output;
 }
 
 export function getCurrentSelectedWidgetConnectionTargets( editor: Editor ) {
@@ -267,64 +182,6 @@ export function focusOnDeskShape( editor: Editor, shapeId: TLShapeId ) {
 	return true;
 }
 
-export function getDeskWidgetConnectionLabel( widget: DeskWidget ) {
-	const props = widget.widgetProps as Record< string, unknown >;
-	switch ( widget.type as string ) {
-		case 'post':
-			return typeof props.postId === 'number' ? `Post #${ props.postId }` : 'Post';
-		case 'page':
-			return typeof props.pageId === 'number' ? `Page #${ props.pageId }` : 'Page';
-		case NOTE_WIDGET_TYPE:
-			return 'Note';
-		case 'site-preview':
-			return typeof props.path === 'string' && props.path ? props.path : 'Preview';
-		case 'site-card':
-			return 'Site card';
-		case 'bookmark':
-		case 'embed':
-			return getUrlHostLabel( props.url ) ?? ( widget.type === 'embed' ? 'Embed' : 'Bookmark' );
-		case 'media':
-			return typeof props.mediaKind === 'string' && props.mediaKind === 'video' ? 'Video' : 'Image';
-		case 'drawing':
-			return 'Drawing';
-		case 'scratchpad':
-			return typeof props.title === 'string' && props.title ? props.title : 'Scratchpad';
-		case 'blog':
-			return 'Blog';
-		case 'post-collection':
-			return 'Posts';
-		default:
-			return widget.type;
-	}
-}
-
-export function getDeskWidgetConnectionTitle(
-	widget: DeskWidget,
-	label = getDeskWidgetConnectionLabel( widget )
-) {
-	const props = widget.widgetProps as Record< string, unknown >;
-	if ( widget.type === 'media' ) {
-		const alt = typeof props.alt === 'string' ? props.alt.trim() : '';
-		return alt || label;
-	}
-
-	if ( widget.type === NOTE_WIDGET_TYPE ) {
-		const text = typeof props.text === 'string' ? getRichTextPlainText( props.text ).trim() : '';
-		return text || label;
-	}
-
-	return label;
-}
-
-export function getDeskWidgetConnectionPillBg( widget: DeskWidget ) {
-	if ( widget.type !== NOTE_WIDGET_TYPE ) {
-		return undefined;
-	}
-
-	const tone = ( widget.widgetProps as { tone?: unknown } ).tone;
-	return NOTE_CONNECTION_PILL_BG[ tone as NoteTone ];
-}
-
 interface DeskConnectorEndpoints {
 	sourceShapeId: TLShapeId;
 	targetShapeId: TLShapeId;
@@ -337,47 +194,6 @@ function getWidgetForShapeId( editor: Editor, shapeId: TLShapeId ) {
 
 function getArrowBindingTerminal( props: object ) {
 	return ( props as { terminal?: unknown } ).terminal;
-}
-
-function getUrlHostLabel( value: unknown ) {
-	if ( typeof value !== 'string' || ! value ) {
-		return null;
-	}
-
-	try {
-		return new URL( value ).hostname.replace( /^www\./, '' );
-	} catch {
-		return value;
-	}
-}
-
-function getRichTextPlainText( value: string ) {
-	if ( typeof document === 'undefined' ) {
-		return stripMarkup( value );
-	}
-
-	return RichTextData.fromHTMLString( value ).toPlainText();
-}
-
-const NOTE_CONNECTION_PILL_BG: Record< NoteTone, string > = {
-	grey: '#6b7280',
-	yellow: '#c4a300',
-	mint: '#3ca56f',
-	blue: '#2271b1',
-	orange: '#c97223',
-	violet: '#7b3fb6',
-	'neon-yellow': '#a18a00',
-	'neon-green': '#2e9e3a',
-	'neon-violet': '#6f2daa',
-	'neon-orange': '#b97917',
-	'neon-blue': '#1873c9',
-};
-
-function stripMarkup( value: string ) {
-	return value
-		.replace( /<[^>]*>/g, ' ' )
-		.replace( /\s+/g, ' ' )
-		.trim();
 }
 
 function isPointInBounds(
