+++ apps/ui/src/ui-desks/desk/selection-toolbar/index.tsx
@@ -4,13 +4,13 @@ import { useEffect, useState } from 'react';
 import { ChatButton } from '@/ui-desks/chats/chat-button';
 import { SelectionChatDialog } from '@/ui-desks/chats/selection-chat-dialog';
 import { Divider, Button, Surface } from '@/ui-desks/components';
-import { appendIncomingConnectedWidgets } from '@/ui-desks/connectors/utils';
+import { appendIncomingConnectedWidgets } from '@/ui-desks/connectors/context';
 import { ControlRenderer } from '@/ui-desks/controls/registry';
 import { useDesk } from '@/ui-desks/desk/provider';
 import { SITE_PREVIEW_WIDGET_TYPE } from '@/ui-desks/widgets/site-preview/types';
 import styles from './style.module.css';
 import type { getSelectedWidgetToolbarItem } from './selection';
-import type { DeskWidgetConnectionTarget } from '@/ui-desks/connectors/utils';
+import type { DeskWidgetConnectionTarget } from '@/ui-desks/connectors/context';
 import type { AnySelectControlConfig } from '@/ui-desks/controls/types';
 import type { StackViewMode } from '@/ui-desks/stacks/utils';
 import type { DeskWidget } from '@/ui-desks/widgets/types';
